fixed cursor positioning by moving cursors container to overlay with game-board
