+++ src/client/package.json
@@ -14,7 +14,6 @@
     "framer-motion": "^11.12.0",
     "lodash": "^4.17.21",
     "lodash.throttle": "^4.1.1",
-    "perfect-cursors": "^1.0.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.x.x",
