+++ src/client/src/components/css/MainBoard.css
@@ -43,12 +43,29 @@
   display: flex;
   gap: 24px;
   align-items: flex-start;
+  position: relative;
 }
 
 .game-board {
   background: #ddd;
   padding: 1px;
+  display: grid;
+  gap: 1px;
   position: relative;
+}
+
+.board-with-cursors {
+  position: relative;
+}
+
+.cursors-container {
+  position: absolute;
+  pointer-events: none;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 10;
 }
 
 .game-status {
